Refactor: split delegate.rs into focused submodules
Decompose the 792-line delegate.rs into three focused files under
delegate/ to satisfy CodeScene 400-line guideline:

- delegate/mod.rs: ChatDelegate struct definition (39 lines)
- delegate/tools.rs: Tool execution helpers (380 lines)
- delegate/loops.rs: NativeLoopDelegate trait implementation (280 lines)

This mirrors the decomposition pattern used for routine_heartbeat.rs
and builtin_tool_coverage.rs.
